pyproject.toml
[project]
name = "sdrbot"
version = "0.1.0-alpha.1"
description = "AI SDR Agent for RevOps"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "deepagents>=0.1.0",
    "langchain-openai>=0.1.0",
    "langchain-anthropic>=0.1.0",
    "langchain-google-genai>=0.1.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "prompt_toolkit>=3.0.0",
    "simple-salesforce>=1.12.0",
    "hubspot-api-client>=8.0.0",
    "keyring>=24.0.0",
    "tavily-python>=0.3.0",
    "requests>=2.31.0",
    "markdownify>=0.11.0",
    "pyinstaller>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-socket>=0.7.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "pre-commit>=3.5.0",
]

[project.scripts]
sdrbot = "sdrbot_cli.main:cli_main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["sdrbot_cli"]

[tool.ruff]
line-length = 100
exclude = []

[tool.ruff.format]
docstring-code-format = true  # Formats code blocks in docstrings

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "F",   # Pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
]
ignore = [
    "E501",    # Line too long
    "COM812",  # Messes with the formatter
    "ISC001",  # Messes with the formatter
    "B008",    # Do not perform function call in argument defaults
]
unfixable = ["B028"]  # Rules that shouldn't be auto-fixed

[tool.ruff.lint.pyupgrade]
keep-runtime-typing = true

[tool.ruff.lint.flake8-annotations]
allow-star-arg-any = true

[tool.ruff.lint.pydocstyle]
convention = "google"  # Google-style docstrings

[tool.ruff.lint.per-file-ignores]
"sdrbot_cli/main.py" = [
    "T201",    # Allow print statements in CLI
]
"tests/*" = [
    "D1",      # Skip documentation rules in tests
    "S101",    # Allow asserts in tests
    "S311",    # Allow pseudo-random generators in tests
    "ANN201",  # Missing return type annotation
    "INP001",  # Implicit namespace package
    "PLR2004", # Magic value comparisons are fine in tests
]

[tool.pytest.ini_options]
timeout = 10  # Default timeout for all tests (can be overridden per-test)
markers = [
    "integration: marks tests that hit real external APIs (deselect with '-m \"not integration\"')",
]

[tool.mypy]
strict = true
ignore_missing_imports = true
enable_error_code = ["deprecated"]
# Optional: reduce strictness if needed
disallow_any_generics = false
warn_return_any = false
